shift most mmime functions to wrapmime
